langchain-project/requirements.txt
openai==1.13.3
langchain==0.1.11
langchain-openai==0.0.8
python-dotenv==1.0.1
faiss-cpu==1.8.0
pypdf==4.1.0
langchainhub==0.1.15
pandas==2.2.2